250412 작업
- 사운드 추가
  -총알 관련 사운드는 deque으로 채널 관리
 - 효과음, 배경음 추가
- 보스 보상 수정, 아이템 생성 움직임 추가

+++ Isaac/Isaac/CMonsterBullet.cpp
@@ -46,7 +46,7 @@ int CMonsterBullet::Update()
 	__super::Update_Rect();
 
 	Set_CollisionBoxPos(m_tInfo.fX - 4.f, m_tInfo.fY);
-	m_tFrame.iStart = m_tStat.fAttack - 1;
+	m_tFrame.iStart = (int)m_tStat.fAttack - 1;
 
 	return NOEVENT;
 }
